pyproject.toml
[project]
name = "matsuzawa-20260213-3"
version = "0.1.0"
description = "FAB-SC experiment: Falsification-Augmented Bidirectional Self-Consistency for LLM reasoning"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "torch>=2.1.0",
    "datasets>=2.14.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]
